chore(jest): separated server tests based on server type
