+++ src/commands/flow/spec/flow.spec.server.viya.ts
@@ -29,8 +29,41 @@ import { build, deploy } from '../..'
 import { execute } from '../execute'
 import SASjs from '@sasjs/adapter/node'
 
-describe('sasjs flow', () => {
-  const target: Target = generateTarget()
+describe('sasjs flow with Viya', () => {
+  dotenv.config()
+  const timestamp = generateTimestamp()
+  const targetName = `cli-tests-flow-${timestamp}`
+
+  const target: Target = new Target({
+    name: targetName,
+    serverType: ServerType.SasViya,
+    serverUrl: process.env.VIYA_SERVER_URL,
+    appLoc: `/Public/app/cli-tests/${targetName}`,
+    contextName,
+    serviceConfig: {
+      serviceFolders: ['sasjs/testServices', 'sasjs/testJob', 'sasjs/services'],
+      initProgram: 'sasjs/testServices/serviceinit.sas',
+      termProgram: 'sasjs/testServices/serviceterm.sas',
+      macroVars: {}
+    },
+    jobConfig: {
+      jobFolders: ['sasjs/testJob'],
+      initProgram: 'sasjs/testServices/serviceinit.sas',
+      termProgram: 'sasjs/testServices/serviceterm.sas',
+      macroVars: {}
+    },
+    authConfig: {
+      client: process.env.CLIENT as string,
+      secret: process.env.SECRET as string,
+      access_token: process.env.ACCESS_TOKEN as string,
+      refresh_token: process.env.REFRESH_TOKEN as string
+    },
+    deployConfig: {
+      deployServicePack: true,
+      deployScripts: []
+    }
+  })
+
   const sasjs: SASjs = new SASjs({
     serverUrl: target.serverUrl,
     httpsAgentOptions: target.httpsAgentOptions,
@@ -45,7 +78,14 @@ describe('sasjs flow', () => {
     authConfig = await getAuthConfig(target)
 
     await createTestApp(__dirname, target.name)
-    await copyJobsAndServices(target.name)
+    await copy(
+      path.join(__dirname, 'testJob'),
+      path.join(__dirname, target.name, 'sasjs', 'testJob')
+    )
+    await copy(
+      path.join(__dirname, 'testServices'),
+      path.join(__dirname, target.name, 'sasjs', 'testServices')
+    )
     await build(target)
     await deploy(target, false)
 
@@ -367,51 +407,3 @@ describe('sasjs flow', () => {
     expect(csvData.match(csvRowRegExp)!.length).toEqual(2)
   })
 })
-
-function generateTarget(serverType = ServerType.SasViya): Target {
-  dotenv.config()
-  const timestamp = generateTimestamp()
-  const targetName = `cli-tests-flow-${timestamp}`
-  return new Target({
-    name: targetName,
-    serverType,
-    serverUrl: (serverType === ServerType.SasViya
-      ? process.env.VIYA_SERVER_URL
-      : process.env.SAS9_SERVER_URL) as string,
-    appLoc: `/Public/app/cli-tests/${targetName}`,
-    contextName,
-    serviceConfig: {
-      serviceFolders: ['sasjs/testServices', 'sasjs/testJob', 'sasjs/services'],
-      initProgram: 'sasjs/testServices/serviceinit.sas',
-      termProgram: 'sasjs/testServices/serviceterm.sas',
-      macroVars: {}
-    },
-    jobConfig: {
-      jobFolders: ['sasjs/testJob'],
-      initProgram: 'sasjs/testServices/serviceinit.sas',
-      termProgram: 'sasjs/testServices/serviceterm.sas',
-      macroVars: {}
-    },
-    authConfig: {
-      client: process.env.CLIENT as string,
-      secret: process.env.SECRET as string,
-      access_token: process.env.ACCESS_TOKEN as string,
-      refresh_token: process.env.REFRESH_TOKEN as string
-    },
-    deployConfig: {
-      deployServicePack: true,
-      deployScripts: []
-    }
-  })
-}
-
-const copyJobsAndServices = async (appName: string) => {
-  await copy(
-    path.join(__dirname, 'testJob'),
-    path.join(__dirname, appName, 'sasjs', 'testJob')
-  )
-  await copy(
-    path.join(__dirname, 'testServices'),
-    path.join(__dirname, appName, 'sasjs', 'testServices')
-  )
-}
